# proc: optimize parseG